Add tests for MOviesHelper and move ListMoviesQuery tests to correct file
